Fix checkbox not being saved

+++ options.php
@@ -158,7 +158,7 @@ class useo_options
             qa_opt('useo_social_schema_enable', (int)qa_post_text('useo_social_schema_enable'));
             qa_opt('useo_social_schema_page_type', (int)qa_post_text('useo_social_schema_page_type'));
             qa_opt('useo_social_gp_thumbnail', qa_post_text('useo_social_gp_thumbnail'));
-            qa_opt('useo_social_enable_editor', qa_post_text('useo_social_enable_editor'));
+            qa_opt('useo_social_enable_editor', (int)qa_post_text('useo_social_enable_editor'));
             // Tags
             qa_opt('useo_tag_desc_max_len', (int)qa_post_text('useo_tag_desc_max_len'));
             qa_opt('useo_tag_desc_enable_icon', (int)qa_post_text('useo_tag_desc_enable_icon'));
